docs(countdown): 调整countdown使用demo，文档站点正常展示示例代码
