+++ src/count-down/_example/style/base.less
@@ -1,4 +1,4 @@
-.t-countdown-type-demo {
+.t-countdown-base-demo {
   padding: 0 16px;
 
   .demo-item {
