+++ src/count-down/_example/style/specs.less
@@ -4,7 +4,7 @@
   .demo-group {
     margin-bottom: 24px;
 
-    &-lebal {
+    &-label {
       color: rgba(0, 0, 0, 0.4);
       font-size: 10px;
       margin-bottom: 17px;
