+++ src/index.ts
@@ -14,3 +14,4 @@ export * from './tabs';
 export * from './switch';
 export * from './sticky';
 export * from './checkbox';
+export * from './count-down';
